refactor: PendingMenuProductsCreator 추출
