+++ src/main/java/kitchenpos/menu/application/MenuService.java
@@ -1,19 +1,11 @@
 package kitchenpos.menu.application;
 
-import static java.util.stream.Collectors.toMap;
-
 import java.util.List;
-import java.util.Map;
-import java.util.stream.Collectors;
-import kitchenpos.menu.domain.repository.MenuGroupRepository;
-import kitchenpos.menu.domain.repository.MenuRepository;
-import kitchenpos.product.domain.repository.ProductRepository;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuGroup;
-import kitchenpos.menu.domain.PendingMenuProduct;
 import kitchenpos.menu.domain.PendingMenuProducts;
-import kitchenpos.product.domain.Product;
-import kitchenpos.menu.ui.request.MenuProductRequest;
+import kitchenpos.menu.domain.repository.MenuGroupRepository;
+import kitchenpos.menu.domain.repository.MenuRepository;
 import kitchenpos.menu.ui.request.MenuRequest;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -23,46 +15,25 @@ public class MenuService {
 
     private final MenuRepository menuRepository;
     private final MenuGroupRepository menuGroupRepository;
-    private final ProductRepository productRepository;
+    private final PendingMenuProductsCreator pendingMenuProductsCreator;
 
     public MenuService(
             final MenuRepository menuRepository,
             final MenuGroupRepository menuGroupRepository,
-            final ProductRepository productRepository
+            final PendingMenuProductsCreator pendingMenuProductsCreator
     ) {
         this.menuRepository = menuRepository;
         this.menuGroupRepository = menuGroupRepository;
-        this.productRepository = productRepository;
+        this.pendingMenuProductsCreator = pendingMenuProductsCreator;
     }
 
     @Transactional
     public Menu create(final MenuRequest request) {
         final MenuGroup menuGroup = menuGroupRepository.findById(request.getMenuGroupId())
                 .orElseThrow(IllegalArgumentException::new);
-        final PendingMenuProducts products = createPendingMenuProducts(request);
+        final PendingMenuProducts products = pendingMenuProductsCreator.create(request);
         final Menu menu = menuGroup.createMenu(request.getName(), request.getPrice(), products);
         return menuRepository.save(menu);
-    }
-
-    private PendingMenuProducts createPendingMenuProducts(final MenuRequest request) {
-        final List<Product> products = findAllProductsByIdIn(request);
-        final Map<Long, Long> quantities = request.getMenuProducts().stream()
-                .collect(toMap(MenuProductRequest::getProductId, MenuProductRequest::getQuantity));
-        final List<PendingMenuProduct> pendingMenuProducts = products.stream()
-                .map(p -> new PendingMenuProduct(p.getId(), p.getPrice(), quantities.get(p.getId())))
-                .collect(Collectors.toList());
-        return new PendingMenuProducts(pendingMenuProducts);
-    }
-
-    private List<Product> findAllProductsByIdIn(final MenuRequest request) {
-        final List<Long> productIds = request.getMenuProducts().stream()
-                .map(MenuProductRequest::getProductId)
-                .collect(Collectors.toList());
-        final List<Product> products = productRepository.findAllByIdIn(productIds);
-        if (products.size() != productIds.size()) {
-            throw new IllegalArgumentException();
-        }
-        return products;
     }
 
     @Transactional
